OCaml 4.07 support
Relevant changes in 4.07
- unix.c now #include <sys/ioctl.h> (since 852b595ff3), but an empty one is fine
  (only use is in caml_nun_rows_fd in an #ifdef TIOCGWINSZ), see
  https://github.com/ocaml/ocaml/pull/1431/
  -> provide an empty ioctl via nolibc
- bigarray is now part of the stdlib
  https://github.com/ocaml/ocaml/pull/1685
  -> libotherlibs.a is no longer needed (neither built nor linked)

To support these changes, an empty sys/ioctl.h is added to nolibc.  Also,
`ocaml-freestanding.pc.in` now contains in `Libs`: -L${libdir} -lasmrum -lnolibc
-lopenlibm (instead of ${libdir}/lib for each library).  `configure.sh` adds
-lotherlibs to these flags in the case that OCAML_GTE_4_07_0 is false.  The
`Makefile` treats libotherlibs.a as dependency only if OCAML_GTE_4_07_0 is false.

+++ install.sh
@@ -40,8 +40,12 @@ else
     done
 fi
 cp build/ocaml/asmrun/libasmrun.a ${DESTLIB}/libasmrun.a
-# OCaml "otherlibs"
-cp build/ocaml/otherlibs/libotherlibs.a ${DESTLIB}/libotherlibs.a
+
+# Prior to OCaml 4.07.0, "otherlibs" contained the bigarray implementation.
+# OCaml >= 4.07.0 includes bigarray as part of stdlib/libasmrun.a
+if [ -f build/ocaml/otherlibs/libotherlibs.a ]; then
+    cp build/ocaml/otherlibs/libotherlibs.a ${DESTLIB}/libotherlibs.a
+fi
 
 # pkg-config
 mkdir -p ${prefix}/lib/pkgconfig
